+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/rule/MasterSlaveRule.java
@@ -20,9 +20,9 @@ package org.apache.shardingsphere.core.rule;
 import com.google.common.base.Preconditions;
 import org.apache.shardingsphere.api.config.masterslave.MasterSlaveDataSourceConfiguration;
 import org.apache.shardingsphere.api.config.masterslave.MasterSlaveRuleConfiguration;
-import org.apache.shardingsphere.core.rule.event.DataSourceNameDisabledEvent;
+import org.apache.shardingsphere.underlying.common.rule.event.impl.DataSourceNameDisabledEvent;
 import org.apache.shardingsphere.underlying.common.rule.DataSourceRoutedRule;
-import org.apache.shardingsphere.underlying.common.rule.RuleChangedEvent;
+import org.apache.shardingsphere.underlying.common.rule.event.RuleChangedEvent;
 import org.apache.shardingsphere.underlying.common.rule.StatusContainedRule;
 
 import java.util.Collection;
